pyproject.toml
[project]
name = "crewai-tiger21"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.193.2",
    "crewai-tools>=0.73.1",
    "google-generativeai>=0.8.5",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.0.10",
    "python-dotenv>=1.1.1",
]
